requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.112.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
Mako==1.3.5
MarkupSafe==2.1.5
openai==1.43.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
regex==2024.7.24
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
